API change: Hooks should register a InvocationHandlerFactory
Instantiating the invocation handler will usually pull referenced classes
which makes them unavailable for instrumentation. The order of mods could
make it impossible  for later mods to instrument a class if it has been
referenced by another mod.

By using the factory the InvocationHandler is created when the hook is
called for the first time. By that time all mods should have loaded and
registered their hooks.