Call chained ActionPerformers in a loop, not recursively
The WrappedBehaviour ensures the actions are called in order.
An ActionPerformer can elect to not run the server action after all
ActionPerformers and is not required to maintain the chain.

It is however possible for an ActionPerformer to call the server action
directly and thus bypassing the propagation stop. This may also result in
duplicate calls to the server action.