Changes to action propagation for standard actions
The default return value for standard actions (those defined by the
server) is now "false". This will prevent unimplemented methods in the
ActionPerformer to always cancel server actions. The default for custom
actions remains "true". Unimplemented methods will stop the action. This
is not an issue since the action is most likely only used by the mod
which defined it.

You can use the propagate() method to set the propagation of the action
to the server and other mods and get the return value for the action method.

#37